move the settings data to router.js
